requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Validation
pydantic>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# LangChain Core and Extensions
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0

# Vector Database
chromadb>=0.4.0
pinecone-client>=2.2.4

# Google AI SDK
google-generativeai>=0.3.0

# Document Processing
pypdf>=3.17.0

# === CODE ANALYSIS DEPENDENCIES ===
# AST Parsing and Code Analysis
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4
tree-sitter-javascript>=0.20.3
tree-sitter-typescript>=0.20.3
tree-sitter-java>=0.20.2
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.4

# Code Processing and Analysis
ast-decompiler>=0.7.0
libcst>=1.1.0
rope>=1.11.0

# Git Integration
GitPython>=3.1.40
pygit2>=1.13.3

# Code Embeddings and Similarity
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0

# Database and Caching
sqlalchemy>=2.0.23
alembic>=1.12.1
redis>=5.0.1
psycopg2-binary>=2.9.9

# Async Processing
celery>=5.3.4
aiofiles>=23.2.1

# Code Quality and Analysis
pylint>=3.0.3
flake8>=6.1.0
mypy>=1.7.1
bandit>=1.7.5

# File Processing
watchdog>=3.0.0
pathspec>=0.11.2

# API Documentation
openapi-spec-validator>=0.7.1

# Additional Dependencies
typing-extensions>=4.8.0
numpy>=1.24.0
pandas>=2.1.0
networkx>=3.2.1 
radon>=5.1.0 